% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle.R
\name{ata}
\alias{ata}
\title{ata}
\usage{
ata(tri, ...)
}
\arguments{
\item{...}{additional arguments}

\item{object}{object to turn into exhibit}
}
\description{
creates age-to-age development triangles
}
\examples{

origin_data <- c(2015, 2016, 2017, 2015, 2016, 2015)
age_data <- c(1, 1, 1, 2, 2, 3)
value_data <- c(10, 11, 10, 15, 16, 17)

my_tri <- triangle(origin = origin_data,
                   age = age_data,
                   value = value_data)

ata(my_tri)

}
