% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle.R
\name{dev_tri}
\alias{dev_tri}
\title{dev_tri}
\usage{
dev_tri(origin, age, value)
}
\arguments{
\item{origin}{origin}

\item{age}{age}

\item{value}{value}
}
\description{
\code{dev_tri} class constructor.
}
\examples{

origin_data <- c(2015, 2016, 2017, 2015, 2016, 2015)
age_data <- c(1, 1, 1, 2, 2, 3)
value_data <- c(10, 11, 10, 15, 16, 17)

my_tri <- dev_tri(origin = origin_data,
              age = age_data,
              value = value_data)

}
