% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development-method.R
\name{development}
\alias{development}
\title{development}
\usage{
development()
}
\arguments{
\item{tri}{}
}
\description{
function for running development reserve estimation method
on an object of class `tidy_tri`
}
\examples{

}
