% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldf.R
\name{idf2cdf}
\alias{idf2cdf}
\title{idf2cdf}
\usage{
idf2cdf(.idf)
}
\arguments{
\item{.idf}{object of class \code{idf}}
}
\description{
convert \link{\code{idf}} object to \link{\code{cdf}} object
}
\examples{

my_idf <- idf(idfs = c(1.75, 1.25, 1.15, 1.1, 1.05, 1.03), first_age = 1)

idf2cdf(.idf = my_idf)
}
