% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle.R
\name{ldf_avg_wtd}
\alias{ldf_avg_wtd}
\title{ldf_avg_wtd}
\usage{
ldf_avg_wtd(tri)
}
\arguments{
\item{tri}{\code{dev_tri} object}
}
\description{
dollar weighted average of eacg age in a \code{dev_tri}
}
\examples{


}
