% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldf.R
\name{plot.idf_}
\alias{plot.idf_}
\title{plot.idf_}
\usage{
\method{plot}{idf_}(idf_)
}
\arguments{
\item{idf_}{object of class \code{idf_}}
}
\description{
generif plot function for class \code{idf_}
}
\examples{

my_idf <- idf(idfs = c(1.75, 1.25, 1.15, 1.1, 1.05, 1.03), first_age = 1)

plot(my_idf)

}
