% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{premium_data}
\alias{premium_data}
\title{Premium data to go alonf with \code{tri_data}}
\format{A data frame with 6 rows and 4 variables:
\describe{
  \item{lob}{Line of Business}
  \item{gr_code}{Unique company code}
  \item{gr_name}{Company name}
  \item{premium_net}{Premium net of ceded reinsurance}
}}
\usage{
premium_data
}
\description{
This data is a subset of the the \code{premium} data frame from the
\code{casdata} package available on Github \code{https://github.com/Ractuary/casdata}
}
\examples{
# code used to create this dataset
if (interactive()) {
  library(casdata)
  library(dplyr)
  library(devtools)

  premium_data <- premium \%>\% filter(lob == "wkcomp", gr_code == 671, origin >= 1992) \%>\%
                                select(lob, gr_code, gr_name, premium_net)

  use_data(premium_data)
}


}
\keyword{datasets}
