% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RAA}
\alias{RAA}
\title{Data from the Reinsurance Association of America (RAA)}
\format{A data frame with 55 rows and 3 variables:
\describe{
  \item{origin}{the origin period of the losses}
  \item{age}{the development age of the losses}
  \item{value}{the value of the losses}
}}
\usage{
RAA
}
\description{
This data was originally provided by the excellent \code{ChainLadder}
and was slightly modified for use in \code{devtri}
}
\keyword{datasets}
