% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tail.R
\name{tail}
\alias{tail}
\title{tail}
\usage{
tail(type, args)
}
\arguments{
\item{type}{type of tail factor}

\item{args}{list of arguments to pass to fit that type of tail}
}
\description{
class for ldf tail factor
}
