% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tri_data}
\alias{tri_data}
\title{Data for making example triangles}
\format{A data frame with 21 rows and 3 variables:
\describe{
  \item{origin}{the origin period of the losses}
  \item{dev}{the development age of the losses}
  \item{value}{the value of the losses}
}}
\usage{
tri_data
}
\description{
Data for making example triangles
}
\keyword{datasets}
