% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldf.R
\name{cdf}
\alias{cdf}
\title{cdf}
\usage{
cdf(cdfs, first_age = 1)
}
\arguments{
\item{cdfs}{the cumulative loss development factors.}

\item{first_age}{the first development age.  This must be a number between 0 and 1.}
}
\description{
Create an object of class \code{cdf}.
}
\examples{
cdf(cdfs = c(2.40, 1.5, 1.2, 1.15, 1.08, 1.03), first_age = 1)
cdf(cdfs = c(2.70, 1.5, 1.2, 1.15, 1.08, 1.03), first_age = 0.5)
}
