% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldf.R
\name{cdf2idf}
\alias{cdf2idf}
\title{idf2cdf}
\usage{
cdf2idf(.cdf)
}
\arguments{
\item{idf}{object of class \code{idf}}
}
\description{
convert \link{\code{idf}} object to \link{\code{cdf}} object
}
\examples{

my_idf <- idf(ldfs = c(1.75, 1.25, 1.15, 1.1, 1.05, 1.03), first_age = 1)

my_cdf <- idf2cdf(idf = .idf)

test <- cdf2idf(my_cdf)
}
