% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldf.R
\name{idf2cdf}
\alias{idf2cdf}
\title{idf2cdf}
\usage{
idf2cdf(idf_)
}
\arguments{
\item{idf_}{object of class \code{idf_}}
}
\description{
convert \code{idf} object to \code{cdf} object
}
\examples{

my_idf <- idf(idfs = c(1.75, 1.25, 1.15, 1.1, 1.05, 1.03), first_age = 1)

idf2cdf(idf_ = my_idf)
}
