% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{interpolate_linear}
\alias{interpolate_linear}
\title{interpolate_linear}
\usage{
interpolate_linear(.cdf, dev = 0.5)
}
\arguments{
\item{.cdf}{cdf object}

\item{dev}{development time to interpolate for}
}
\description{
interpolate a \code{cdf} object to future ages
}
\examples{

my_cdf <- cdf(ldfs = c(2.70, 1.5, 1.2, 1.15, 1.08, 1.03), first_age = 0.5)

interpolate_linear(my_idf, dev = 0.75)
}
